requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
openpyxl==3.1.2
pandas==2.1.3
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.8.0
click==8.1.7
python-multipart
